% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivariatenormcheck.R
\name{bivariatenormcheck}
\alias{bivariatenormcheck}
\title{bivariatenormcheck}
\usage{
bivariatenormcheck(x1, x2)
}
\value{
Text output describing result of the test
}
\description{
Uses a bivariate normality check to calculate the proportion of bivariate data falling in an expected 50% confidence ellipse
}
\details{
User provides two data vectors for function to check the pairs of data against 50% confidence ellipse
}
\examples{
bivariatenormcheck(x1 = c(1,4,3,3,5,6,7), x2 = c(20,15,8,7,40,30,22))
}
