% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bonfcisvariance.R
\name{bonfcisvariance}
\alias{bonfcisvariance}
\title{bonfcisvariance}
\usage{
bonfcisvariance(data, q = 1, alpha = 0.05)
}
\arguments{
\item{data}{the data set to calculate the intervals for the variances}

\item{q}{the number of variances to estimate}

\item{alpha}{the confidence level for the intervals}
}
\value{
Bonferonni confidence intervals for the first q principal component variances
}
\description{
Produces Bonferroni confidence intervals for the first q principal components of a sample
}
\examples{
\dontrun{ bonfcisvariance(data = MATH4793MOSER::stocks, q = 2, alpha = 0.1)}
}
