% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisquaredqq.R
\name{chisquaredqq}
\alias{chisquaredqq}
\title{chisqauredqq}
\usage{
chisquaredqq(data)
}
\arguments{
\item{data}{the data set to compute the chi squared quantiles for}
}
\value{
list output with theoretical chi squared quantiles and sorted calculated quantiles
}
\description{
Produces sorted data and theoretical quantiles for a chi-squared QQ plot for a provided data set
}
\details{
The quantiles and sorted data can be used by a plotting pacakge to create the QQ plot
}
\examples{
chisquaredqq(MATH4793MOSER::welder)
}
