% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidenceellipse.R
\name{confidenceellipse}
\alias{confidenceellipse}
\title{confidenceellipse}
\usage{
confidenceellipse(x1, x2, alpha = 0.05)
}
\arguments{
\item{x1}{the first varaible's data}

\item{x2}{the second varaible's data}

\item{alpha}{the confidence level for the confidence ellipse}
}
\value{
A ggplot object containing the scatterplot with confidence ellipse overlayed
}
\description{
Generates a confidence ellipse for two variables at a given confidence level
}
\details{
User provides two data vectors to use as variables in question and alpha level
}
\examples{
data <- MATH4793MOSER::welder
\dontrun{ confidenceellipse(data$voltage,data$current,alpha=0.1)}
}
