% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotellingsttest.R
\name{hotellingsttest}
\alias{hotellingsttest}
\title{hotellingsttest}
\usage{
hotellingsttest(data, mu, alpha)
}
\value{
List output of calculations from Hotelling's T Test and a plot of the confidence ellipse plot
}
\description{
This function performs calculations and makes a plot for a bivariate dataset based on Hotelling's T test
}
\details{
User inputs a bivariate dataset, estimate for the mu vector, and alpha value corresponding to the desired confidence level and gets the outcome of the test, information about the data provided, and a plot of the confidence ellipse
}
\section{List of outputs includes}{

- Outcome of the test
- The calculated quadratic size
- The scaled quantile
- The eigenvectors of the variance covariance matrix
- The eigenvalues of the variance covariance matrix
- The half major and minor axis lengths of the confidence ellipse
- The ratio of the major and minor axes in for the confidence ellipse
}

\section{The plot}{

The plot shows the confidence ellipse with lines marking the mean values and the axes of the ellipse
}

\examples{
hotellingsttest(data = matrix(c(1,2,3,4,5,6),nrow=3,ncol=2), mu = matrix(c(1,2),nrow=2, ncol=1), alpha=0.05)
}
