% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvnormalityapp.R
\name{mvnormalityapp}
\alias{mvnormalityapp}
\title{mvnormalityapp}
\usage{
mvnormalityapp()
}
\value{
Normal proportion test outcomes, violin plots, standard normal and chi squared QQ plots, a confidence ellipse
}
\description{
basic shiny dashboard that generates some output about multivariate nromality checks
}
\details{
User has ability to change plot views with dynamic inputs
}
\examples{
\dontrun{ mvnormalityapp()}
}
