% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcabootAG.R
\name{pcabootAG}
\alias{pcabootAG}
\title{pcabootAG}
\usage{
pcabootAG(data, alpha = 0.05, iter = 1000, cov = FALSE)
}
\arguments{
\item{data}{the data set to run the principal component analysis on}

\item{alpha}{decimal value to express the desired 100(1-alpha) percent confidence level}

\item{iter}{the number of iterations to be used in the bootstrap sampling}

\item{cov}{True or false value to indicate use of covariance or correlation matrix where true uses the covariance matrix}
}
\value{
the function automatically plots density histograms for each eigenvalue and a boxplot of estimates for the first two principal components (eigenvectors). It also returns bootstrap confidence intervals for all eigenvector components and eigenvalues as well as Anderson Girshick theoretical confidence intervals for the eigenvalues. The intervals are listed in order on the command line for each eigenvalue. Components are listed by vector in order.
}
\description{
this function produces confidence intervals for the eigenvector components and eigenvalues of a given data set
}
\details{
This function produces confidence intervals using bootstrap methods as well as theoretical large sample results from Anderson and Girshick
}
\examples{
\dontrun{ pcabootAG(MATH4793MOSER::welder, alpha = 0.15, iter = 2000, cov = TRUE)}
}
