% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/principalcomponents.R
\name{principalcomponents}
\alias{principalcomponents}
\title{principalcomponents}
\usage{
principalcomponents(data)
}
\arguments{
\item{data}{data set for which some principal component calculations are to be performed}
}
\value{
A list containing the sample covariance matrix, the principal components, the eigenvalues, the total variance, and a matrix of the correlations between the variables and the principal components
}
\description{
Computes relevant values and output related to the principal components of a sample of data
}
\details{
Computes the covariance matrix for the sample, the principal components, the lambda values, the total variance and the correlations between the variables and principal components
}
\examples{
\dontrun{ principalcomponents(data = MATH4793MOSER::welder)}
}
