% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsquaredchart.R
\name{tsquaredchart}
\alias{tsquaredchart}
\title{tsquaredchart}
\usage{
tsquaredchart(data, conflevel1 = 0.95, conflevel2 = 0.99)
}
\arguments{
\item{data}{the data set to create the plot for}

\item{conflevel1}{first confidence level for the chart. Decimal represents percentage NOT alpha level}

\item{conflevel2}{second confidence level for the chart. Decimal representes percentage NOT alpha level}
}
\value{
A ggplot object containing the plot
}
\description{
Creates a t squared plot for a data set
}
\details{
The t squared chart helps the suer see abnormal observations based on 95 and 99% confidence threshholds
}
\examples{
\dontrun{ tsquaredchart(data = MATH4793MOSER::welder)}
}
