% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariateqq.R
\name{univariateqq}
\alias{univariateqq}
\title{univariateqq}
\usage{
univariateqq(x)
}
\arguments{
\item{x}{the data vector}
}
\value{
List ouput containing the quantiles and sorted data
}
\description{
Produces the quantiles and sorted data to produce a standard normal QQ plot for a single variable
}
\details{
User provides data as a vector and the quantiles and sorted data are returned
}
\examples{
univariateqq(x = c(2,4,3,7,9,1,3,5,6))
}
