% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rq.R
\name{rq}
\alias{rq}
\title{rq}
\usage{
rq(x)
}
\arguments{
\item{x}{the vector of data to calculate rq for}
}
\value{
List ouput containing the rq value
}
\description{
Calculates the correlation coefficient between sorted data and theoretical quantiles in a univariate QQ plot
}
\details{
For a variable, univariateqq() is invoked to get quantiles, then the corelation coefficient is calculated
}
\examples{
rq(x = c(2,4,3,7,9,1,3,5,6))
}
