% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsquaredchart.R
\name{tsquaredchart}
\alias{tsquaredchart}
\title{tsquaredchart}
\usage{
tsquaredchart(data)
}
\arguments{
\item{data}{the data set to create the plot for}
}
\value{
A ggplot object containing the plot
}
\description{
Creates a t squared plot for a data set
}
\details{
The t squared chart helps the suer see abnormal observations based on 95 and 99% confidence threshholds
}
\examples{
\dontrun{ tsquaredchart(data = MATH4793MOSER::welder)}
}
